+++ test/testBasicManiProduct.jl
@@ -328,7 +328,7 @@ end
     )
 
     @test !isPartial(P12)
-    @test P12._partial === nothing
+    @test getPartial(P12) === nothing
     # @test isapprox( mean(P12)[1], 0, atol=1 )
     # @test isapprox( mean(P12)[2], 0, atol=1 )